Split PointNucleus to a separate file
