+++ src/CMakeLists.txt
@@ -12,6 +12,7 @@ general/dftfuncs.cpp diatomic/basis.cpp diatomic/quadrature.cpp
 diatomic/dftgrid.cpp diatomic/twodquadrature.cpp
 general/model_potential.cpp
 general/ModelPotential.cpp
+general/PointNucleus.cpp
 general/RadialPotential.cpp
 general/GaussianNucleus.cpp
 )
